refactor(extension): rename modules/ → lib/ to avoid WXT name collision
WXT reserves `modules/` for its plugin system. Our directory held plain
ES modules (content-script + background helpers), which caused WXT's
file watcher to classify every save under `modules/` as a WXT-plugin
change and trigger a full dev-server + browser restart.

The previous commit papered over this with a `modulesDir: '.wxt-modules'`
override. This commit does the idiomatic thing: rename our directory to
`lib/` so the override is no longer needed.

- 20 pure file renames: packages/extension/modules/** → .../lib/**
- entrypoints/background.js: 5 imports
- entrypoints/content/index.js: 9 imports
- entrypoints/bridge.content.js: 1 comment reference
- wxt.config.js: modulesDir override removed

`wxt build` succeeds, bundle sizes unchanged.
