+++ packages/extension/lib/background/api-sync.js
