+++ packages/extension/lib/background/badge.js
