+++ packages/extension/lib/background/export.js
