+++ packages/extension/lib/background/utils.js
