+++ packages/extension/lib/content/annotation-popover.js
