+++ packages/extension/lib/content/api-bridge.js
