+++ packages/extension/lib/content/badge-manager.js
