+++ packages/extension/lib/content/bridge-handler.js
