+++ packages/extension/lib/content/element-context.js
