+++ packages/extension/lib/content/event-bus.js
