+++ packages/extension/lib/content/floating-toolbar.js
