+++ packages/extension/lib/content/inspection-mode.js
