+++ packages/extension/lib/content/orchestrator.js.pre-wxt
