+++ packages/extension/lib/content/popover-panels.js
