+++ packages/extension/lib/content/shadow-dom-utils.js
