+++ packages/extension/lib/content/shadow-host.js
