+++ packages/extension/lib/content/styles.js
