+++ packages/extension/lib/content/theme-manager.js
