+++ packages/extension/lib/content/toolbar-docs.js
